removes packages.ini  usage
makes thigns complex and not WinPython business
