+++ winpython/__init__.py
@@ -28,6 +28,6 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
 """
 
-__version__ = '8.2.20240618'
+__version__ = '9.0.20240623'
 __license__ = __doc__
 __project_url__ = 'http://winpython.github.io/'
