+++ winpython/utils.py
@@ -23,6 +23,7 @@ import sys
 import stat
 import locale
 import io
+import configparser as cp
 
 # Local imports
 import winreg
@@ -860,6 +861,59 @@ def formatted_list(list_of_list, full=False, max_width=70):
             for line in list_of_list
         ]
         return zz
+
+# pep503 defines normalized package names: www.python.org/dev/peps/pep-0503
+def normalize(name):
+    """return normalized (unique) name of a package"""
+    return re.sub(r"[-_.]+", "-", name).lower()
+
+def get_package_metadata(database, name, update=False, suggested_summary=None):
+    """Extract infos (description, url) from the local database"""
+    # for package.ini safety belt
+    # Note: we could use the PyPI database but this has been written on
+    # machine which is not connected to the internet
+    # we store only  normalized names now (PEP 503)
+    DATA_PATH = str(Path(sys.modules['winpython'].__file__).parent /'data')
+    db = cp.ConfigParser()
+    filepath = Path(database) if Path(database).is_absolute() else Path(DATA_PATH) / database
+    try:
+        db.read_file(open(str(filepath), encoding = 'utf-8'))
+    except:
+        db.read_file(open(str(filepath)))
+    my_metadata = dict(
+        description="",
+        url="https://pypi.org/project/" + name,
+    )
+    for key in my_metadata:
+        # wheel replace '-' per '_' in key
+        for name2 in (name, normalize(name)):
+            try:
+                my_metadata[key] = db.get(name2, key)
+                break
+            except (cp.NoSectionError, cp.NoOptionError):
+                pass
+    db_desc = my_metadata.get("description")
+
+    if my_metadata.get("description") == "" and suggested_summary:
+        # nothing in package.ini, we look in our installed packages
+        try:
+            my_metadata["description"] = (
+                suggested_summary + "\n"
+            ).splitlines()[0]
+        except:
+            pass
+
+    if update == True and db_desc == "" and my_metadata["description"] != "":
+        # we add new findings in our packgages.ini list, if it's required
+        try:
+            db[normalize(name)] = {}
+            db[normalize(name)]["description"] = my_metadata["description"]
+            with open(str(Path(DATA_PATH) / database), "w",  encoding='UTF-8') as configfile:
+                db.write(configfile)
+        except:
+            pass
+    return my_metadata
+
 
 if __name__ == '__main__':
